requirements.txt
aiosmtpd==1.4.5
atpublic==4.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.7
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.7